Starting with assignmnet  315F - NumberSlidingGame
Class for game initialization is implemented but not done SlideBoardGame).

Class for datamodel (board) is implemented (SlideBoardModel)
Class for acting on the datamodel is done via a third party object (SlideBoardActions).
Some implementations are yet to be done on this class.
Most importanly the sliding behaviour (moving number to open tile.